+++ src/helpers/Url.php
@@ -9,16 +9,9 @@ use yii\helpers\BaseUrl;
 
 class Url extends BaseUrl
 {
-    public static function draft(string $url): string
+    public static function draft(array|string $route): string
     {
-        $request = Yii::$app->getRequest();
-
-        if ($request->draftSubdomain === false) {
-            return $url;
-        }
-
-        return !$request->getIsDraft() || !str_contains($url, (string) $request->draftSubdomain)
-            ? preg_replace('#^((https?://)(www.)?)#', "$2$request->draftSubdomain.", $url)
-            : $url;
+        $url = static::to($route);
+        return Yii::$app->getUrlManager()->createDraftUrl($url);
     }
 }