+++ src/views/migration.php
@@ -8,7 +8,8 @@ declare(strict_types=1);
  * @var string $namespace the new migration class namespace
  */
 
-echo "<?php\n";
+echo "<?php\n\ndeclare(strict_types=1);\n";
+
 if (!empty($namespace)) {
     echo "\nnamespace $namespace;\n";
 }